fix: MovieDetailResponseSchema and base_url in pagination

+++ src/schemas/movies.py
@@ -9,15 +9,15 @@ class MovieDetailResponseSchema(BaseModel):
     id: int
     name: str
     date: datetime.date
-    score: int
+    score: float
     genre: str
     overview: str
     crew: str
     orig_title: str
     status: str
     orig_lang: str
-    budget: float | None = None
-    revenue: float | None = None
+    budget: int | None = None
+    revenue: int | None = None
     country: str
